migrations: fix paradata copy in the migration
While the migrations worked when it was developed, when migrating
existing old surveys, the logs to paradata migration hangs when
inserting data. The paused stream seems to be causing deadlock. Instead
of using a stream, we simply insert paradata one interview at a time, as
this operation is not time-critical.

incidentally fixes #991 as the inner query is not there anymore